Refactor placeable method structure
