+++ src/main/java/net/quickwrite/fluent4j/util/StringSliceUtil.java
@@ -1,14 +1,22 @@
 package net.quickwrite.fluent4j.util;
 
+import net.quickwrite.fluent4j.ast.FluentAttribute;
 import net.quickwrite.fluent4j.ast.FluentElement;
+import net.quickwrite.fluent4j.ast.FluentVariant;
 import net.quickwrite.fluent4j.ast.placeable.*;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentPlaceable;
+import net.quickwrite.fluent4j.ast.placeable.base.FluentSelectable;
 import net.quickwrite.fluent4j.exception.FluentParseException;
+import net.quickwrite.fluent4j.exception.FluentSelectException;
+import net.quickwrite.fluent4j.parser.FluentParser;
 import net.quickwrite.fluent4j.util.args.FluentArgs;
 import net.quickwrite.fluent4j.util.args.FunctionFluentArgs;
 import net.quickwrite.fluent4j.util.args.FunctionFluentArguments;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * A collection of different utility methods
@@ -227,6 +235,50 @@ public final class StringSliceUtil {
         return expression;
     }
 
+    /**
+     * Returns a simple {@link FluentArgs} object that
+     * contains the arguments that can be created in the
+     * {@code .ftl}-files.
+     *
+     * <p>
+     * The arguments are always in a format
+     * of being in {@code ()} and being separated
+     * by {@code ,}.
+     * </p>
+     * <p>
+     * There are two different types of arguments:
+     *     <ul>
+     *         <li>
+     *             Positional Arguments:
+     *             <br />
+     *             Arguments that don't have a name and
+     *             are just denominated by their position in the
+     *             arguments.
+     *         </li>
+     *         <li>
+     *             Named Arguments:
+     *             <br />
+     *             Arguments that have a name that is being
+     *             declared in front of the argument by
+     *             having a {@code :} as a separator.
+     *             <br />
+     *             This means that to create a named argument
+     *             the argument is created by having this format:
+     *             <code>{{name}} : {{argument}}</code>
+     *             <br />
+     *             They can only be accessed by their name as
+     *             the key.
+     *         </li>
+     *     </ul>
+     * </p>
+     * <p>
+     *     If there are no arguments in the brackets
+     *     the function will return {@link FluentArgs#EMPTY_ARGS}.
+     * </p>
+     *
+     * @param content The content the arguments are in
+     * @return The arguments as an option
+     */
     private static FluentArgs parseArguments(final StringSlice content) {
         if (content == null) {
             return FluentArgs.EMPTY_ARGS;
@@ -321,5 +373,135 @@ public final class StringSliceUtil {
         }
 
         return slice.substring(start, slice.getPosition());
+    }
+
+    /**
+     * Returns the parsed version of the {@link FluentPlaceable}
+     * that is inside the <code>{}</code>.
+     *
+     * <p>
+     * This can either be a simple expression or
+     * a complete {@link SelectExpression}.
+     * </p>
+     *
+     * @param content The content that the placeable currently is at
+     * @return The Placeable
+     */
+    public static FluentPlaceable getPlaceable(final StringSlice content) {
+        StringSliceUtil.skipWhitespaceAndNL(content);
+
+        FluentPlaceable placeable = StringSliceUtil.getExpression(content);
+
+        boolean canSelect = placeable instanceof FluentSelectable;
+
+        StringSliceUtil.skipWhitespaceAndNL(content);
+
+        if (canSelect && content.getChar() == '-') {
+            content.increment();
+            if (content.getChar() != '>') {
+                throw new FluentParseException("->", "-" + content.getCharUTF16(), content.getAbsolutePosition());
+            }
+
+            content.increment();
+
+            StringSliceUtil.skipWhitespaceAndNL(content);
+
+            final List<FluentVariant> fluentVariants = new ArrayList<>();
+            FluentVariant defaultVariant = null;
+
+            while (content.getChar() != '}') {
+                final Pair<FluentVariant, Boolean> variant = getVariant(content);
+
+                fluentVariants.add(variant.getLeft());
+
+                if (!variant.getRight()) {
+                    continue;
+                }
+
+                if (defaultVariant != null) {
+                    throw new FluentSelectException("Only one variant can be marked as default (*)");
+                }
+
+                defaultVariant = variant.getLeft();
+            }
+
+            if (defaultVariant == null) {
+                throw new FluentSelectException("Expected one of the variants to be marked as default (*)");
+            }
+
+            placeable = new SelectExpression(placeable, fluentVariants, defaultVariant);
+        }
+
+        StringSliceUtil.skipWhitespaceAndNL(content);
+
+        return placeable;
+    }
+
+    private static Pair<FluentVariant, Boolean> getVariant(final StringSlice content) {
+        StringSliceUtil.skipWhitespaceAndNL(content);
+
+        boolean isDefault = false;
+
+        if (content.getChar() == '*') {
+            isDefault = true;
+            content.increment();
+        }
+
+        if (content.getChar() != '[') {
+            throw new FluentParseException('[', content.getCharUTF16(), content.getAbsolutePosition());
+        }
+
+        content.increment();
+
+        StringSliceUtil.skipWhitespace(content);
+
+        final StringSlice identifier = getVariantIdentifier(content);
+
+        StringSliceUtil.skipWhitespace(content);
+
+        if (content.getChar() != ']') {
+            throw getVariantException(content, identifier.toString(), "]");
+        }
+
+        content.increment();
+
+        final Pair<StringSlice, Integer> stringSliceContent = FluentParser.getMessageContent(content,
+                character -> character == '[' || character == '*' || character == '}');
+
+        final FluentAttribute attribute = new FluentAttribute(
+                identifier,
+                stringSliceContent.getLeft(),
+                stringSliceContent.getRight()
+        );
+
+        return new ImmutablePair<>(new FluentVariant(attribute), isDefault);
+    }
+
+    private static StringSlice getVariantIdentifier(final StringSlice content) {
+        char character = content.getChar();
+        final int start = content.getPosition();
+
+        while (character != ' '
+                && character != '\n'
+                && character != ']'
+                && character != '\0'
+        ) {
+            content.increment();
+            character = content.getChar();
+        }
+
+        return content.substring(start, content.getPosition());
+    }
+
+    private static FluentParseException getVariantException(final StringSlice content, final String prev, final String expected) {
+        int start = content.getPosition();
+
+        while (content.getChar() != ']' && !content.isBigger()) {
+            content.increment();
+        }
+
+        return new FluentParseException(expected,
+                prev + content.substring(start, content.getPosition()).toString(),
+                content.getAbsolutePosition());
     }
 }
